plantDetection/huggingface/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
tflite-runtime==2.14.0
numpy==1.24.3
Pillow==10.4.0
python-multipart==0.0.9
